adding selecting metric computation options and SVD computation

+++ checkpoint_metric_computation.py
@@ -11,6 +11,7 @@ import pickle
 import os 
 import time 
 from tqdm import tqdm
+from scipy.linalg import svdvals
 
 cpu_count = os.cpu_count()
 
@@ -118,6 +119,7 @@ def compute_grad_weight_magnitudes(gradient_dataset):
     for checkpoint_step in tqdm(checkpoint_steps):
 
         print("Processing checkpoint step: ", checkpoint_step)
+
         checkpoint_step_idx_range = checkpoint_step_to_range_indices[checkpoint_step]
 
         checkpoint_grads = gradient_dataset.select(range(*checkpoint_step_idx_range))
@@ -153,7 +155,7 @@ def compute_grad_weight_magnitudes(gradient_dataset):
     return grad_weight_magnitudes_per_layer
 
 
-def compute_grad_sim_per_layer(gradient_dataset):
+def compute_grad_sim(gradient_dataset):
     """
     Computes the cosine similarity of the gradients in each layer of the model at the different timesteps.
     """
@@ -168,6 +170,7 @@ def compute_grad_sim_per_layer(gradient_dataset):
     for checkpoint_step in tqdm(checkpoint_steps):
 
         print("Processing checkpoint step: ", checkpoint_step)
+
         checkpoint_step_idx_range = checkpoint_step_to_range_indices[checkpoint_step]
 
         checkpoint_grads = gradient_dataset.select(range(*checkpoint_step_idx_range))
@@ -213,6 +216,65 @@ def compute_grad_sim_per_layer(gradient_dataset):
     
     return grad_sim_per_layer
 
+def compute_svd(dataset):
+    """
+    For each layer, we compute the singular values.
+    """
+
+    print("Computing SVD directions per layer")
+
+    layer_names = set(dataset['layer_name'])
+    # NOTE: we are assuming we are extracting 3 projections per layer
+    num_layers = len(layer_names)//3 
+
+    svd_weight_per_layer = dict()
+
+    checkpoint_step_to_range_indices = get_checkpoint_step_to_range_indices(dataset)
+
+    for checkpoint_step in tqdm(checkpoint_steps):
+
+        print("Processing checkpoint step: ", checkpoint_step)
+
+        checkpoint_step_idx_range = checkpoint_step_to_range_indices[checkpoint_step]
+
+        checkpoint_weights = dataset.select(range(*checkpoint_step_idx_range))
+
+        # NOTE: Doing this because I want to be sure that we are extracting the correct matrices
+        # (since the actual filter operation is slow)
+        layer_name_to_indices = {}
+        for idx, layer_name in enumerate(checkpoint_weights['layer_name']):
+            if layer_name not in layer_name_to_indices:
+                layer_name_to_indices[layer_name] = []
+            layer_name_to_indices[layer_name].append(idx)
+        
+        # computing the OV projection 
+        for layer_num in range(num_layers):
+
+            qkv_idx = layer_name_to_indices[f"gpt_neox.layers.{layer_num}.attention.query_key_value"]
+            output_idx = layer_name_to_indices[f"gpt_neox.layers.{layer_num}.attention.dense"]
+
+            qkv_projection = np.array(checkpoint_weights.select(qkv_idx)['data'][0])
+            value_projection = qkv_projection[-qkv_projection.shape[0]//3:,]
+            output_projection = np.array(checkpoint_weights.select(output_idx)['data'][0])
+
+            # these are the two matrices we are interested in
+            ov_projection =  output_projection @ value_projection 
+            mlp_dense = np.array(checkpoint_weights.select(layer_name_to_indices[f"gpt_neox.layers.{layer_num}.mlp.dense_4h_to_h"])['data'][0])
+
+            layer_projection_matrices = dict()
+            layer_projection_matrices[f"gpt_neox.layers.{layer_num}.attention.ov_projection"] = ov_projection   
+            layer_projection_matrices[f"gpt_neox.layers.{layer_num}.mlp.dense_4h_to_h"] = mlp_dense
+
+            for layer_name, proj_matrix in layer_projection_matrices.items():
+
+                if layer_name not in svd_weight_per_layer:
+                    svd_weight_per_layer[layer_name] = []
+
+                S = svdvals(proj_matrix)
+                svd_weight_per_layer[layer_name].append(S)
+
+    return svd_weight_per_layer
+
 def get_dataset(subconfig: str):
     retry_count = 0
     sleep_time = 10
@@ -221,7 +283,7 @@ def get_dataset(subconfig: str):
         try: 
             dataset = load_dataset(
                 "rdiehlmartinez/pythia-training-metrics", subconfig, split='default',
-                cache_dir='/rds-d7/user/rd654/hpc-work/cache',
+                # cache_dir='/rds-d7/user/rd654/hpc-work/cache',
                 writer_batch_size=100,
             )
             break
@@ -238,39 +300,70 @@ def get_dataset(subconfig: str):
 
 
 @click.command()
-@click.option('--model_size') 
-def main(model_size):
+@click.argument('model_size') 
+@click.option(
+    '--metrics', '-m', multiple=True,
+    default=[
+        "cka",
+        "weight_magnitudes", 
+        "grad_weight_magnitudes", 
+        "grad_sim",
+        "svd_weight", 
+        "svd_grad_weight"
+    ])
+def main(model_size, metrics):
 
-    activation_dataset = get_dataset(f"{model_size}__activations")
-    weights_dataset = get_dataset(f"{model_size}__weights")
-    gradient_dataset = get_dataset(f"{model_size}__gradients_mini")
+    activation_dataset = None
+    weights_dataset = None
+    gradient_dataset = None
 
     # save out the computed metrics to compiled_satistics/model_size
     os.makedirs(f"compiled_statistics/{model_size}", exist_ok=True)
  
     cka_scores_per_layer_fn = f"compiled_statistics/{model_size}/cka_scores_per_layer.pkl"
-    if not os.path.exists(cka_scores_per_layer_fn):
+    if not os.path.exists(cka_scores_per_layer_fn) and "cka" in metrics:
+        if activation_dataset is None:
+            activation_dataset = get_dataset(f"{model_size}__activations")
+
         cka_scores_per_layer = compute_cka_sores(activation_dataset)
         with open(cka_scores_per_layer_fn, "wb") as f:
             pickle.dump(cka_scores_per_layer, f)
 
     weight_magnitudes_per_layer_fn = f"compiled_statistics/{model_size}/weight_magnitudes_per_layer.pkl"
-    if not os.path.exists(weight_magnitudes_per_layer_fn):
+    if not os.path.exists(weight_magnitudes_per_layer_fn) and "weight_magnitudes" in metrics:
+        if weights_dataset is None:
+            weights_dataset = get_dataset(f"{model_size}__weights")
+
         weight_magnitudes_per_layer = compute_weight_magnitudes(weights_dataset)
         with open(weight_magnitudes_per_layer_fn, "wb") as f:
             pickle.dump(weight_magnitudes_per_layer, f)
 
     grad_weight_magnitudes_per_layer_fn = f"compiled_statistics/{model_size}/grad_weight_magnitudes_per_layer.pkl"
-    if not os.path.exists(grad_weight_magnitudes_per_layer_fn):
+    if not os.path.exists(grad_weight_magnitudes_per_layer_fn) and "grad_weight_magnitudes" in metrics:
+        if gradient_dataset is None:
+            gradient_dataset = get_dataset(f"{model_size}__gradients")
+
         grad_weight_magnitudes_per_layer = compute_grad_weight_magnitudes(gradient_dataset)
         with open(grad_weight_magnitudes_per_layer_fn, "wb") as f:
             pickle.dump(grad_weight_magnitudes_per_layer, f)
 
     grad_sim_per_layer_fn = f"compiled_statistics/{model_size}/grad_sim_per_layer.pkl"
-    if not os.path.exists(grad_sim_per_layer_fn):
-        grad_sim_per_layer = compute_grad_sim_per_layer(gradient_dataset)
+    if not os.path.exists(grad_sim_per_layer_fn) and "grad_sim" in metrics:
+        if gradient_dataset is None:
+            gradient_dataset = get_dataset(f"{model_size}__gradients")
+
+        grad_sim_per_layer = compute_grad_sim(gradient_dataset)
         with open(grad_sim_per_layer_fn, "wb") as f:
             pickle.dump(grad_sim_per_layer, f)
+
+    svd_weight_per_layer_fn = f"compiled_statistics/{model_size}/svd_weight_per_layer.pkl"
+    if not os.path.exists(svd_weight_per_layer_fn) and "svd_weight" in metrics:
+        if weights_dataset is None:
+            weights_dataset = get_dataset(f"{model_size}__weights")
+
+        svd_weight_per_layer = compute_svd(weights_dataset)
+        with open(svd_weight_per_layer_fn, "wb") as f:
+            pickle.dump(svd_weight_per_layer, f) 
 
     
 if __name__ == '__main__':
